+++ integration/chat-plugin/websocket_test.spec.ts
@@ -1,5 +1,17 @@
 import {cyChatPluginMessageList} from 'chat-plugin-handles';
 
+// declare global {
+//   namespace Cypress {
+//     interface Chainable<Subject> {
+//       createChatPluginConversation(): Chainable<Subject>;
+//       editContactDetails(): Chainable<Subject>;
+//       connectChatPluginConnector(): Chainable<Subject>;
+//       editDisplayName(): Chainable<Subject>;
+//       listConversationsForContact(): Chainable<Subject>;
+//     }
+//   }
+// }
+
 describe('Websocket test', () => {
   it('Send message from Inbox to Chatplugin and assert Websocket is working', () => {
     cy.createChatPluginConversation();
