--- integration/tsconfigCy.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "lib": ["es5", "dom"],
-    "types": ["cypress", "node", "./support"]
-  },
-  "include": [
-    "**/*.ts"
-  ]
-}
