+++ integration/ui/connect_chatplugin.spec.ts
@@ -1,5 +1,17 @@
 import {cyChannelsFormBackButton, cyChannelsChatPluginList} from 'handles';
 
+// declare global {
+//   namespace Cypress {
+//     interface Chainable<Subject> {
+//       createChatPluginConversation(): Chainable<Subject>;
+//       editContactDetails(): Chainable<Subject>;
+//       connectChatPluginConnector(): Chainable<Subject>;
+//       editDisplayName(): Chainable<Subject>;
+//       listConversationsForContact(): Chainable<Subject>;
+//     }
+//   }
+// }
+
 describe('Connect chatplugin channel', () => {
   it('Connect chatplugin channel', () => {
     cy.visit('/control-center/connectors');
